LAST COMMIT

+++ FITHUB WEB API/Controllers/GenericController.cs
@@ -1,9 +1,10 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using FHB_BLL.DTO;
-using FHB_BLL.Services;
 using FHB_BLL.Services.GenericServices;
 using FHB_BLL.Wrapping;
 using FHB_BLL.Wrapping.Exceptions;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 namespace FITHUB_WEB_API.Controllers
 {
@@ -18,27 +19,26 @@ namespace FITHUB_WEB_API.Controllers
             _service = service;
         }
 
-        [HttpGet("GetAll")]
-        public ApiResponse<IEnumerable<Dto>> GetAll()
-        {
-            return _service.GetAll();
-        }
-
-        [HttpGet("GetById")]
-        public ApiResponse<Dto> GetById(int id)
-        {
-            return _service.GetById(id);
-        }
-
+       
         [HttpPost("Add")]
         public ApiResponse<Dto> Add(Dto dto)
         {
+            if (dto == null)
+            {
+                throw new ValidationException("DTO is not null");
+            }
+
             return _service.Add(dto);
         }
 
         [HttpPut("Update")]
         public ApiResponse<Dto> Update(Dto dto)
         {
+            if (dto == null)
+            {
+                throw new ValidationException("DTO cannot be null");
+            }
+
             return _service.Update(dto);
         }
 
@@ -51,7 +51,29 @@ namespace FITHUB_WEB_API.Controllers
         [HttpDelete("Delete")]
         public ApiResponse<bool> Delete(Dto dto)
         {
+            if (dto == null)
+            {
+                throw new ValidationException("DTO cannot be null");
+            }
+
             return _service.Delete(dto);
         }
+
+
+        [HttpGet("GetAll")]
+        public ApiResponse<IEnumerable<Dto>> GetAll()
+        {
+
+
+
+            return _service.GetAll();
+        }
+
+        [HttpGet("GetById")]
+        public ApiResponse<Dto> GetById(int id)
+        {
+            return _service.GetById(id);
+        }
+
     }
 }
